+++ packages/bpk-component-spinner/src/BpkExtraLargeSpinner-test.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import { colorBlue500 } from 'bpk-tokens/tokens/base.es6';
 import BpkExtraLargeSpinner from './BpkExtraLargeSpinner';
 
 describe('BpkExtraLargeSpinner', () => {
@@ -10,8 +9,8 @@ describe('BpkExtraLargeSpinner', () => {
     expect(tree).toMatchSnapshot();
   });
 
-  it('should render correctly with a "fill" attribute', () => {
-    const tree = renderer.create(<BpkExtraLargeSpinner fill={colorBlue500} />).toJSON();
+  it('should render correctly with a "className" attribute', () => {
+    const tree = renderer.create(<BpkExtraLargeSpinner className="my-custom-class" />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 });